add OData wrapper for all MS API calls - RootElement deprecated.
Mediasite API calls have odata properties at the top level JSON data,
along with the value property containing the data asked for.
Was using RestClient RootElement to process Value directly and ignore odata.
RootElement is deprecated so we are adding an OData wrapper to capture
these properties on every call, and strip them out when returning the data.